+++ fxn/types/__init__.py
@@ -5,5 +5,5 @@
 
 from .dtype import Dtype
 from .prediction import Acceleration, Prediction, PredictionResource
-from .predictor import AccessMode, EnumerationMember, Parameter, Predictor, PredictorStatus, Signature
+from .predictor import EnumerationMember, Parameter, Predictor, PredictorAccess, PredictorStatus, Signature
 from .user import User